[functions] Add clearNodeParent(), clearNodeTransform(), getNodeScene()
