+++ packages/functions/src/index.ts
@@ -51,6 +51,8 @@
  */
 
 export * from './center';
+export * from './clear-node-parent';
+export * from './clear-node-transform';
 export * from './colorspace';
 export * from './dedup';
 export * from './dequantize';
@@ -72,6 +74,7 @@ export * from './texture-compress';
 export * from './tangents';
 export * from './texture-resize';
 export * from './transform-mesh';
+export * from './transform-primitive';
 export * from './unlit';
 export * from './unpartition';
 export * from './unweld';
